fix: Use custom USDC bridge in WorldChain_SpokePool (#647)
* fix: Support OP bridged USDC in WorldChain adapter


---------

+++ contracts/test/MockBedrockStandardBridge.sol
@@ -20,7 +20,7 @@ contract MockBedrockL2StandardBridge is IL2ERC20Bridge {
         address _remoteToken,
         address _to,
         uint256 _amount,
-        uint256 _minGasLimit,
+        uint32 _minGasLimit,
         bytes calldata _extraData
     ) external {
         // Check that caller has approved this contract to pull funds, mirroring mainnet's behavior
